+++ src/main/java/me/pajic/affogato_core/config/ModClientConfig.java
@@ -1,5 +1,6 @@
 package me.pajic.affogato_core.config;
 
+import me.fzzyhmstrs.fzzy_config.annotations.NonSync;
 import me.fzzyhmstrs.fzzy_config.annotations.Version;
 import me.fzzyhmstrs.fzzy_config.config.Config;
 import me.fzzyhmstrs.fzzy_config.validation.collection.ValidatedMap;
@@ -7,7 +8,6 @@ import me.fzzyhmstrs.fzzy_config.validation.minecraft.ValidatedIdentifier;
 import me.fzzyhmstrs.fzzy_config.validation.misc.ValidatedBoolean;
 import me.fzzyhmstrs.fzzy_config.validation.number.ValidatedInt;
 import me.pajic.affogato_core.ClientMain;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.Identifier;
 
 import java.util.Map;
@@ -22,12 +22,8 @@ public class ModClientConfig extends Config {
     public ValidatedBoolean raiseHotbar = new ValidatedBoolean();
     public ValidatedInt hotbarRaisePixels = new ValidatedInt(2, Integer.MAX_VALUE, 1);
     public ValidatedBoolean disableBetterGrassLayerFeature = new ValidatedBoolean();
-    public ValidatedMap<Identifier, Integer> perDimensionBrightness = (new ValidatedMap.Builder())
-            .keyHandler(ValidatedIdentifier.ofRegistryKey(
-                    Identifier.withDefaultNamespace("overworld"),
-                    Registries.DIMENSION,
-                    (_, _) -> true
-            ))
+    @NonSync public ValidatedMap<Identifier, Integer> perDimensionBrightness = (new ValidatedMap.Builder())
+            .keyHandler(new ValidatedIdentifier(Identifier.withDefaultNamespace("overworld")))
             .valueHandler(new ValidatedInt(25, 100, 0))
             .defaults(Map.of(
                     Identifier.withDefaultNamespace("overworld"), 25,